Add initial project files and configurations
This commit adds the initial setup files and configurations for the project. It includes .gitignore and .eslintrc.cjs files for GIT and ESLint respectively, TypeScript configuration in package.json, as well as Vite and React setup files. It also includes basic style files and code for the initial version of the app.